# service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for JustPayd services",
        "title": "JustPayd API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/shift_requests": {
            "get": {
                "description": "Admin gets all shift requests, can filter by status, user_id, and shift_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift-requests"
                ],
                "summary": "List all shift requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (pending, approved, rejected)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by shift ID",
                        "name": "shift_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved shift requests",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/shift_requests.ShiftRequestResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not an admin",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "User with role \"user\" creates shift request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift-requests"
                ],
                "summary": "User creates shift request",
                "parameters": [
                    {
                        "description": "Shift request creation payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shift_requests.CreateShiftRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Shift request created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/shift_requests.ShiftRequestResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/shift_requests/approve/{id}": {
            "put": {
                "description": "Admin approves shift request by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift-requests"
                ],
                "summary": "Admin approves shift request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shift Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shift request approved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/shift_requests.ShiftRequestResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not an admin",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Shift request not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/shift_requests/reject/{id}": {
            "put": {
                "description": "Admin rejects shift request by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift-requests"
                ],
                "summary": "Admin rejects shift request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shift Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shift request rejected successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/shift_requests.ShiftRequestResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not an admin",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Shift request not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/shifts": {
            "get": {
                "description": "List semua shift",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "List semua shift",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved shifts",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/shifts.ShiftResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Admin creates shift",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "Admin creates shift",
                "parameters": [
                    {
                        "description": "Shift creation payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shifts.CreateShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Shift created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/shifts.ShiftResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/shifts/{id}": {
            "get": {
                "description": "Mendapatkan detail shift berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "Detail shift",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shift ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved shift detail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/shifts.ShiftResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid shift ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Shift not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Mengedit shift berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "Edit shift",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shift ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Shift update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shifts.UpdateShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shift updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/shifts.ShiftResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or shift ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Shift not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus shift berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "Hapus shift",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shift ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shift deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid shift ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Shift not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pkg.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "shift_requests.CreateShiftRequestDTO": {
            "type": "object",
            "required": [
                "shift_id"
            ],
            "properties": {
                "shift_id": {
                    "type": "integer"
                }
            }
        },
        "shift_requests.ShiftRequestResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "requested_at": {
                    "type": "string"
                },
                "shift_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "shifts.CreateShiftRequest": {
            "type": "object",
            "required": [
                "date",
                "end_time",
                "role",
                "start_time"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "shifts.ShiftResponse": {
            "type": "object",
            "properties": {
                "assignee": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_assigned": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "shifts.UpdateShiftRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}